Fix some bugs caught by warnings.
This code had several cases of trying to use memset(3) to zero out an
array, but was using sizeof(char*) instead of the real size of the
buffer, so I've fixed it up to just use calloc(3) instead.

There were very obvious warnings for this. Along the lines of
"warning: this doesn't do what you think it does". To avoid this kind
of thing going unnoticed in the future, I've turned on -Werror and
-Wall in the project.

Change-Id: I06057976f1f4721dcf9acdd37f2cb8a64b7865a7